moved game to lib folder
