+++ spec/game_spec.rb
@@ -1,5 +1,5 @@
 require "rspec"
-require_relative '../game.rb'
+require_relative '../lib/game.rb'
 
 describe "Game" do 
    it "passes this example test" do
